+++ app/scripts/global-filter/global-filter-service.ts
@@ -1,99 +1,67 @@
 /**
- * @file "Global filters" are similar to token conditions (like `word = "rock"`), but are managed separately in the GUI
- *   and then merged with the tokens of the query when sending it to the backend.
  * @format
  */
-import angular, { IQService } from "angular"
+import angular, { ITimeoutService } from "angular"
 import _ from "lodash"
 import settings from "@/settings"
 import { regescape } from "@/util"
 import { RootScope } from "@/root-scope.types"
 import { LocationService } from "@/urlparams"
 import { countAttrValues } from "@/backend/attr-values"
-import { DataObject, GlobalFilterService, UpdateScope } from "./types"
-import { CqpQuery, Condition } from "@/cqp_parser/cqp.types"
 import { RecursiveRecord } from "@/backend/types/attr-values"
 
+export type GlobalFilterService = {
+    initialize: () => void
+}
+
+/** Shape of data stored in URL */
 type StoredFilterValues = Record<string, string[]>
 
-// Data service for the global filter in korp
-// Directive is duplicated in simple and extended search
-// so this directive holds all state concering the users input
-// and what possible filters and values are available
-
-// diretives calls registerScope to register for updates
-// service calls scope.update() when changes occur
+/**
+ * "Global filters" are text-level CQP conditions for selected attributes, that are managed separately in the GUI and
+ * then merged with the tokens of the query when sending it to the backend. This service manages the state of the
+ * filters, which lives in the root scope. The GUI component is duplicated in simple and extended search, and it
+ * manages user-selected values.
+ */
 angular.module("korpApp").factory("globalFilterService", [
-    "$rootScope",
     "$location",
-    "$q",
-    function ($rootScope: RootScope, $location: LocationService, $q: IQService): GlobalFilterService {
-        const scopes: UpdateScope[] = []
-
-        const notify = () => listenerDef.promise.then(() => scopes.map((scope) => scope.update(dataObj)))
-
-        // deferred for waiting for all directives to register
-        var listenerDef = $q.defer<never>()
-
-        /** Model of filter data. */
-        const dataObj: DataObject = {
-            filterValues: {},
-            defaultFilters: [],
-            attributes: {},
-            showDirective: false,
-        }
-
+    "$rootScope",
+    "$timeout",
+    function ($location: LocationService, $rootScope: RootScope, $timeout: ITimeoutService): GlobalFilterService {
         /** Drilldown of values available for each attr. N-dimensional map where N = number of attrs. */
         let currentData: RecursiveRecord<number> = {}
 
-        /** Populate `filterValues` from `defaultFilters`. */
-        function initFilters() {
-            // delete any filter values that are not in the selected filters
-            for (const filter in dataObj.filterValues) {
-                if (!dataObj.defaultFilters.includes(filter)) {
-                    delete dataObj.filterValues[filter]
-                }
-            }
-
-            // create object for every filter that is selected but not yet created
-            for (const filter of dataObj.defaultFilters) {
-                if (!(filter in dataObj.filterValues)) {
-                    dataObj.filterValues[filter] = { value: [], possibleValues: [] }
-                }
-            }
-        }
-
-        function getSupportedCorpora() {
-            const corporaPerFilter = _.map(dataObj.defaultFilters, (filter) => dataObj.attributes[filter].corpora)
-            return _.intersection(...(corporaPerFilter || []))
-        }
-
         /** Fetch token counts keyed in multiple dimensions by the values of attributes */
         async function getData(): Promise<void> {
-            const corpora = getSupportedCorpora()
-            const attrs = dataObj.defaultFilters
-            const multiAttrs = attrs.filter((attr) => dataObj.attributes[attr].settings.type === "set")
-            currentData = await countAttrValues(corpora, attrs, multiAttrs)
-            updateData()
+            const corpora = settings.corpusListing.getSelectedCorpora()
+            const attrs = Object.keys($rootScope.globalFilterData)
+            const multiAttrs = attrs.filter((attr) => $rootScope.globalFilterData[attr].attribute.type === "set")
+            currentData = corpora.length && attrs.length ? await countAttrValues(corpora, attrs, multiAttrs) : {}
+            $timeout(() => {
+                updateData()
+                // Deselect values that are not in the options
+                for (const filter of Object.values($rootScope.globalFilterData)) {
+                    filter.value = filter.value.filter((value) => filter.options.some(([v]) => v === value))
+                }
+            })
         }
 
         // when user selects an attribute, update all possible filter values and counts
         function updateData() {
             function collectAndSum(
-                filters: string[],
+                attrs: string[],
                 elements: RecursiveRecord<number>,
                 parentSelected: boolean
             ): [number, boolean] {
-                const filter = filters[0]
-                const { possibleValues } = dataObj.filterValues[filter]
-                const currentValues = dataObj.filterValues[filter].value
+                const attr = attrs[0]
+                const filter = $rootScope.globalFilterData[attr]
                 let sum = 0
                 const values: string[] = []
                 let include = false
                 for (let value in elements) {
                     var childCount: number
                     const child = elements[value]
-                    const selected = currentValues.includes(value) || _.isEmpty(currentValues)
+                    const selected = !filter.value.length || filter.value.includes(value)
 
                     // filter of any parent values that do not support the child values
                     include = include || selected
@@ -102,14 +70,12 @@ angular.module("korpApp").factory("globalFilterService", [
                         childCount = child
                         include = true
                     } else {
-                        ;[childCount, include] = collectAndSum(_.tail(filters), child, parentSelected && selected)
+                        ;[childCount, include] = collectAndSum(_.tail(attrs), child, parentSelected && selected)
                     }
 
-                    if (include && parentSelected) {
-                        possibleValues.push([value, childCount])
-                    } else {
-                        possibleValues.push([value, 0])
-                    }
+                    const countDisplay = include && parentSelected ? childCount : 0
+                    filter.options.push([value, countDisplay])
+
                     if (selected && include) {
                         sum += childCount
                     }
@@ -121,145 +87,112 @@ angular.module("korpApp").factory("globalFilterService", [
             }
 
             // reset all filters
-            for (const filter of dataObj.defaultFilters) {
-                dataObj.filterValues[filter].possibleValues = []
-            }
+            for (const filter of Object.values($rootScope.globalFilterData)) filter.options = []
 
             // recursively decide the counts of all values
-            collectAndSum(dataObj.defaultFilters, currentData, true)
+            collectAndSum(Object.keys($rootScope.globalFilterData), currentData, true)
 
             // merge duplicate child values
-            for (const filter of dataObj.defaultFilters) {
-                const possibleValuesTmp: Record<string, number> = {}
-                for (const [value, count] of dataObj.filterValues[filter].possibleValues) {
-                    if (!(value in possibleValuesTmp)) {
-                        possibleValuesTmp[value] = 0
-                    }
-                    possibleValuesTmp[value] += count
+            for (const filter of Object.values($rootScope.globalFilterData)) {
+                // Sum the counts of duplicate values
+                const options: Record<string, number> = {}
+                for (const [value, count] of filter.options) {
+                    options[value] ??= 0
+                    options[value] += count
                 }
-                dataObj.filterValues[filter].possibleValues = []
-                for (let k in possibleValuesTmp) {
-                    const v = possibleValuesTmp[k]
-                    dataObj.filterValues[filter].possibleValues.push([k, v])
-                }
-
-                dataObj.filterValues[filter].possibleValues.sort((a, b) => a[0].localeCompare(b[0], $rootScope.lang))
+                // Cast back to list and sort alphabetically
+                filter.options = Object.entries(options).sort((a, b) => a[0].localeCompare(b[0], $rootScope.lang))
             }
         }
 
         /** Parse encoded url param value to local data. */
         function setFromLocation(globalFilter?: string) {
-            if (!globalFilter) {
-                return
-            }
-            if (!dataObj.filterValues) {
-                return
-            }
+            if (!globalFilter) return
             const parsedFilter: StoredFilterValues = JSON.parse(atob(globalFilter))
 
-            // Set values from param, if corresponding filter is available.
-            for (const attrKey in parsedFilter) {
-                const attrValues = parsedFilter[attrKey]
-                if (dataObj.defaultFilters.includes(attrKey)) {
-                    dataObj.filterValues[attrKey].value = attrValues
-                }
+            // Copy values from param, reset filters not in param
+            for (const attr in $rootScope.globalFilterData) {
+                $rootScope.globalFilterData[attr].value = parsedFilter[attr] || []
             }
-
-            // Set other available filters to empty.
-            for (const attrKey in dataObj.filterValues) {
-                if (!(attrKey in parsedFilter)) {
-                    dataObj.filterValues[attrKey].value = []
-                }
-            }
-        }
-
-        /** Build a CQP token object of AND-combined conditions from active filters. */
-        function makeCqp(): CqpQuery {
-            const andArray: Condition[][] = []
-            for (const attrKey in dataObj.filterValues) {
-                const attrValues = dataObj.filterValues[attrKey]
-                const attrType = dataObj.attributes[attrKey].settings.type
-                andArray.push(
-                    attrValues.value.map((attrValue) => ({
-                        type: `_.${attrKey}`,
-                        op: attrType === "set" ? "contains" : "=",
-                        val: regescape(attrValue),
-                    }))
-                )
-            }
-
-            return [{ and_block: andArray }]
         }
 
         /** Set url param from local data, as base64-encoded json. */
         function updateLocation() {
             const rep: StoredFilterValues = {}
-            for (const attrKey in dataObj.filterValues) {
-                const attrValues = dataObj.filterValues[attrKey]
-                if (!_.isEmpty(attrValues.value)) {
-                    rep[attrKey] = attrValues.value
-                }
-            }
+            Object.entries($rootScope.globalFilterData).forEach(([attr, filter]) => {
+                if (filter.value.length) rep[attr] = filter.value
+            })
             if (!_.isEmpty(rep)) {
                 $location.search("global_filter", btoa(JSON.stringify(rep)))
-                $rootScope.globalFilter = makeCqp()
+                // Build a CQP token object of AND-combined conditions from active filters.
+                $rootScope.globalFilter = [
+                    {
+                        and_block: Object.entries($rootScope.globalFilterData).map(([attr, filter]) =>
+                            filter.value.map((value) => ({
+                                type: `_.${attr}`,
+                                op: filter.attribute.type === "set" ? "contains" : "=",
+                                val: regescape(value),
+                            }))
+                        ),
+                    },
+                ]
             } else {
                 $location.search("global_filter", null)
                 $rootScope.globalFilter = null
             }
         }
 
-        /** Update available filters when changing corpus selection. */
-        $rootScope.$on("corpuschooserchange", () => {
-            if (settings.corpusListing.selected.length === 0) {
-                dataObj.showDirective = false
-            } else {
-                const filterAttributes = settings.corpusListing.getDefaultFilters()
+        function initialize() {
+            /** Update available filters when changing corpus selection. */
+            $rootScope.$on("corpuschooserchange", () => {
+                if (settings.corpusListing.selected.length > 0) {
+                    const attrs = settings.corpusListing.getDefaultFilters()
 
-                // Disable the filters feature if none are applicable to all selected corpora.
-                if (_.isEmpty(filterAttributes)) {
-                    dataObj.showDirective = false
-                    $location.search("global_filter", null)
-                    $rootScope.globalFilter = null
-                    // Unset any active filters.
-                    for (let filter of dataObj.defaultFilters) {
-                        dataObj.filterValues[filter].value = []
+                    // Remove filters that are no more applicable
+                    for (const attr in $rootScope.globalFilterData) {
+                        if (!attrs[attr]) delete $rootScope.globalFilterData[attr]
                     }
-                } else {
-                    dataObj.showDirective = true
-                    dataObj.defaultFilters = Object.keys(filterAttributes)
-                    dataObj.attributes = filterAttributes
 
-                    initFilters()
+                    // Add new filters
+                    for (const attr in attrs) {
+                        $rootScope.globalFilterData[attr] ??= {
+                            attribute: attrs[attr],
+                            value: [], // Selection empty by default
+                            options: [], // Filled in updateData
+                        }
+                    }
 
                     setFromLocation($location.search().global_filter)
                     getData()
                     updateLocation()
-                    notify()
                 }
-            }
-            // Flag that the filter feature is ready.
-            $rootScope.globalFilterDef.resolve()
-        })
+                // Flag that the filter feature is ready.
+                $rootScope.globalFilterDef.resolve()
+            })
 
-        /** Set up sync from url params to local data. */
-        $rootScope.$watch(
-            () => $location.search().global_filter,
-            (filter) => setFromLocation(filter)
-        )
+            /** Set up sync from url params to local data. */
+            $rootScope.$watch(
+                () => $location.search().global_filter,
+                (filter) => setFromLocation(filter)
+            )
+
+            $rootScope.$watch(
+                "globalFilterData",
+                (filterData: RootScope["globalFilterData"], filterDataOld?: RootScope["globalFilterData"]) => {
+                    // Only watch selection changes
+                    const values = _.mapValues(filterData, (filter) => filter.value)
+                    const valuesOld = _.mapValues(filterDataOld, (filter) => filter.value)
+                    if (!_.isEqual(values, valuesOld)) {
+                        updateLocation()
+                        getData()
+                    }
+                },
+                true
+            )
+        }
 
         return {
-            registerScope(scope) {
-                scopes.push(scope)
-                // TODO this will not work with parallel mode since only one directive is used :(
-                if (scopes.length === 2) {
-                    listenerDef.resolve()
-                }
-            },
-            valueChange() {
-                updateLocation()
-                updateData()
-            },
+            initialize,
         }
     },
 ])