+++ app/scripts/global-filter/global-filters.ts
@@ -1,29 +1,28 @@
 /** @format */
-import angular, { IController } from "angular"
+import angular, { IController, IScope } from "angular"
 import _ from "lodash"
 import { html } from "@/util"
 import "./global-filter-service"
-import { DataObject, UpdateScope } from "./types"
 import "./global-filter"
+import { RootScope } from "@/root-scope.types"
+import { GlobalFilterService } from "./global-filter-service"
 
-type GlobalFiltersController = IController & {
-    lang: string
-}
-
-type GlobalFiltersScope = UpdateScope & {
-    dataObj: DataObject
+type GlobalFiltersScope = IScope & {
+    setSelected: (attr: string, selected: string[]) => void
+    show: boolean
 }
 
 angular.module("korpApp").component("globalFilters", {
-    template: html`<div ng-if="dataObj.showDirective" class="mb-4">
+    template: html`<div ng-if="show" class="mb-4">
         <span class="font-bold"> {{ 'global_filter' | loc:$root.lang}}:</span>
         <div class="inline-block">
-            <span ng-repeat="filterKey in dataObj.defaultFilters">
+            <span ng-repeat="(attr, filter) in $root.globalFilterData">
                 <global-filter
-                    attr="filterKey"
-                    attr-def="dataObj.attributes[filterKey]"
-                    attr-value="dataObj.filterValues[filterKey].value"
-                    possible-values="dataObj.filterValues[filterKey].possibleValues"
+                    attr="attr"
+                    attr-def="filter.attribute"
+                    attr-value="filter.value"
+                    options="filter.options"
+                    on-change="setSelected(attr, selected)"
                 ></global-filter>
                 <span ng-if="!$last">{{"and" | loc:$root.lang}}</span>
             </span>
@@ -31,22 +30,26 @@ angular.module("korpApp").component("globalFilters", {
     </div>`,
     bindings: {},
     controller: [
+        "$rootScope",
         "$scope",
         "globalFilterService",
-        function ($scope: GlobalFiltersScope, globalFilterService) {
-            const $ctrl = this as GlobalFiltersController
-
+        function ($rootScope: RootScope, $scope: GlobalFiltersScope, globalFilterService: GlobalFilterService) {
+            const $ctrl = this as IController
             $ctrl.$onInit = () => {
-                globalFilterService.registerScope($scope)
-                $scope.dataObj = {
-                    filterValues: {},
-                    defaultFilters: [],
-                    attributes: {},
-                    showDirective: false,
-                }
+                globalFilterService.initialize()
             }
 
-            $scope.update = (dataObj) => ($scope.dataObj = dataObj)
+            $scope.setSelected = (attr, selected) => {
+                $rootScope.globalFilterData[attr].value = selected
+            }
+
+            $rootScope.$watch(
+                "globalFilterData",
+                () => {
+                    $scope.show = Object.keys($rootScope.globalFilterData).length > 0
+                },
+                true
+            )
         },
     ],
 })